Modularize FRAMEWORKS constant
for convenience when adding new examples
